\name{test.ci-package}
\alias{test.ci-package}
\alias{test.ci}
\docType{package}
\title{
\packageTitle{test.ci}
}
\description{
\packageDescription{test.ci}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{test.ci}
\packageIndices{test.ci}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{test.ci}

Maintainer: \packageMaintainer{test.ci}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
